engine: the pixel index does the arithmetic its own constructor documents
`RgbaImage::new` spends four lines on why `width * height * 4` has to be computed
in `u64` — "an 8192² export is 268 MB and a caller may ask for more, where it wraps
at 4 GB". `RgbaImage::pixel`, eighteen lines below it, then did `((y * width + x) *
4)` entirely in `u32`.

Unreachable when that comment was written and not now: `max_export_dim` is
`max_texture_dimension_2d` specifically so the ceiling can rise with the adapter,
and the adapters this runs on commonly report 32768. At 32768² the multiply crosses
`u32::MAX` in the last rows — a wrapped index reading the wrong pixel in release, an
overflow panic in a getter in debug.

+++ crates/stark-engine/src/image.rs
@@ -30,7 +30,11 @@ impl RgbaImage {
 
     /// The RGBA bytes at `(x, y)`. Panics if out of bounds.
     pub fn pixel(&self, x: u32, y: u32) -> [u8; 4] {
-        let i = ((y * self.width + x) * 4) as usize;
+        // In `u64` like [`RgbaImage::new`]'s check, and for its reason: at the export
+        // ceiling `max_export_dim` now reports (`max_texture_dimension_2d`, commonly
+        // 32768), `y * width + x` crosses `u32::MAX` in the last rows — so the index
+        // this returns would wrap to a pixel somewhere else in the image.
+        let i = (u64::from(y) * u64::from(self.width) + u64::from(x)) as usize * 4;
         [
             self.pixels[i],
             self.pixels[i + 1],
